Fixed zsh bootstrap commands
The current boostrap command for zsh has a bug in it. The newline intended to be appended to the end of the current .zshrc file is instead added to (or creates) a file named 'a' in the execution directory. Which causes problems when the comment is added to the user's .zshrc file if it didn't already end in a newline. 

This fixes that by making the zsh command consistent with the other shells